project_files/requirements.txt
-e .

aws-cdk.core
aws-cdk.aws-dynamodb
aws-cdk.aws-events
aws-cdk.aws-lambda
aws-cdk.aws-s3
aws-cdk.aws-cloudwatch
